requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.35
alembic==1.13.1
motor==3.3.2
pymongo==4.6.1
beanie==1.24.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic>=2.0.0
pydantic-settings==2.1.0

# RAG & LLM
groq==0.4.1
chromadb==0.4.22

# Document Processing
pypdf==4.0.1
python-docx==1.1.0
openpyxl==3.1.2

# Translation (Hindi/English)
langdetect==1.0.9

# Utilities
httpx==0.26.0
aiofiles==23.2.1
python-dateutil==2.8.2
